+++ test/dummy/config/initializers/recording_studio.rb
@@ -1,7 +1,13 @@
 # frozen_string_literal: true
 
 RecordingStudio.configure do |config|
-  config.recordable_types = [ "Workspace", "Folder", "Page" ]
+  config.recordable_types = [
+    "Workspace",
+    "Folder",
+    "Page",
+    "AdminRoot",
+    "RecordingStudioPublishable::Publishable"
+  ]
 
   config.require_recordable_declarations = true